Address plan-managed sheet review
Remount the sheet content per item so a usage draft never carries over when
switching between plan-managed controls, reject non-finite usage before
enabling save, and return one plan usage limit per identity so the
dashboard payload matches the row the section renders.
